new changes

+++ app/Http/Traits/Observable.php
@@ -35,15 +35,17 @@ trait Observable
 
     public static function logChange(Model $model, string $action)
     {
-        if ($action == 'CREATED') {
-            DB::table($model->getTable())->where('id', $model->id)->delete();
-        } elseif ($action == 'UPDATED') {
-            $arr = array_diff_key($model->getOriginal(), ['id' => 'aa']);
-            DB::table($model->getTable())->where('id', $model->id)->update($arr);
-        } elseif ($action == 'DELETED') {
-            DB::table($model->getTable())->insert($model->getOriginal());
+        if (!checkIfSuperAdmin()) {
+            if ($action == 'CREATED') {
+                DB::table($model->getTable())->where('id', $model->id)->delete();
+            } elseif ($action == 'UPDATED') {
+                $arr = array_diff_key($model->getOriginal(), ['id' => 'aa']);
+                DB::table($model->getTable())->where('id', $model->id)->update($arr);
+            } elseif ($action == 'DELETED') {
+                DB::table($model->getTable())->insert($model->getOriginal());
+            }
         }
-       $log =  Log::create([
+        $log = Log::create([
             'user_id' => Auth::user()->id ?? null,
             'model' => static::class,
             'model_id' => $model->id,
@@ -59,10 +61,11 @@ trait Observable
             ])
         ]);
 
-        Notification::addNotification(Notification::APPROVAL_REQUEST, 1);
-        Task::addTask(Notification::APPROVAL_REQUEST, 1,$log->id,$action.' '.$model->getTable(). ' entry');
-
-        Session::put('approvalRequest', 1);
+        if (!checkIfSuperAdmin()) {
+            Notification::addNotification(Notification::APPROVAL_REQUEST, 1);
+            Task::addTask(Notification::APPROVAL_REQUEST, 1, $log->id, $action . ' ' . $model->getTable() . ' entry');
+            Session::put('approvalRequest', 1);
+        }
     }
 
     /**
